requirements.txt
discord
pickle5
google-api-python-client